Changed PassHash internal type
I was running into problems using the crypto library's PasswordString
type inside PassHash. These came to a head when trying to write
QuickCheck tests. I was going to have to write more code to instance
this into Show.

Instead, I decided to simplify the type to ByteString and write explicit
conversion functions to go to/from PasswordString as needed.

This made the PersistField and PersistFieldSql instancing much easier as
well.